Commonize some to jvmMain
