+++ algorithm/src/jvmNonAndroidMain/kotlin/com/alexvanyo/composelife/model/SerializationMessages.jvmNonAndroid.kt
