+++ build-logic/convention/src/main/kotlin/com/alexvanyo/composelife/buildlogic/Molecule.kt
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2024 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alexvanyo.composelife.buildlogic
+
+import org.gradle.api.Project
+import org.gradle.api.attributes.java.TargetJvmEnvironment
+import org.gradle.api.attributes.java.TargetJvmEnvironment.TARGET_JVM_ENVIRONMENT_ATTRIBUTE
+import org.gradle.kotlin.dsl.get
+import org.jetbrains.kotlin.gradle.InternalKotlinGradlePluginApi
+import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
+import org.jetbrains.kotlin.gradle.targets.jvm.KotlinJvmTarget
+
+fun Project.jvmMolecule(kotlinMultiplatformExtension: KotlinMultiplatformExtension): KotlinJvmTarget =
+    with(kotlinMultiplatformExtension) {
+        jvm("molecule") {
+            attributes {
+                attribute(
+                    TARGET_JVM_ENVIRONMENT_ATTRIBUTE,
+                    objects.named(TargetJvmEnvironment::class.java, "molecule"),
+                )
+            }
+            @OptIn(InternalKotlinGradlePluginApi::class)
+            configurations[resourcesElementsConfigurationName].attributes {
+                attribute(
+                    TARGET_JVM_ENVIRONMENT_ATTRIBUTE,
+                    objects.named(TargetJvmEnvironment::class.java, "molecule"),
+                )
+            }
+        }
+    }
