+++ ui-app/src/jbMain/kotlin/com/alexvanyo/composelife/ui/app/action/settings/CellShapeConfigUi.kt
@@ -21,48 +21,25 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.key
-import androidx.compose.runtime.mutableFloatStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.saveable.listSaver
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.alexvanyo.composelife.parameterizedstring.ParameterizedString
 import com.alexvanyo.composelife.parameterizedstring.parameterizedStringResolver
 import com.alexvanyo.composelife.parameterizedstring.parameterizedStringResource
-import com.alexvanyo.composelife.preferences.CurrentShape
-import com.alexvanyo.composelife.preferences.CurrentShapeType
 import com.alexvanyo.composelife.preferences.di.ComposeLifePreferencesProvider
 import com.alexvanyo.composelife.preferences.di.LoadedComposeLifePreferencesProvider
-import com.alexvanyo.composelife.preferences.setCurrentShapeType
-import com.alexvanyo.composelife.preferences.setRoundRectangleConfig
-import com.alexvanyo.composelife.sessionvalue.SessionValue
-import com.alexvanyo.composelife.sessionvalue.localSessionId
-import com.alexvanyo.composelife.sessionvalue.rememberSessionValueHolder
-import com.alexvanyo.composelife.ui.app.component.DropdownOption
 import com.alexvanyo.composelife.ui.app.component.EditableSlider
 import com.alexvanyo.composelife.ui.app.component.IdentitySliderBijection
 import com.alexvanyo.composelife.ui.app.component.TextFieldDropdown
 import com.alexvanyo.composelife.ui.app.resources.CornerFractionLabel
 import com.alexvanyo.composelife.ui.app.resources.CornerFractionLabelAndValue
 import com.alexvanyo.composelife.ui.app.resources.CornerFractionValue
-import com.alexvanyo.composelife.ui.app.resources.RoundRectangle
 import com.alexvanyo.composelife.ui.app.resources.Shape
 import com.alexvanyo.composelife.ui.app.resources.SizeFractionLabel
 import com.alexvanyo.composelife.ui.app.resources.SizeFractionLabelAndValue
 import com.alexvanyo.composelife.ui.app.resources.SizeFractionValue
 import com.alexvanyo.composelife.ui.app.resources.Strings
-import com.alexvanyo.composelife.ui.util.uuidSaver
-import com.benasher44.uuid.uuid4
-import com.livefront.sealedenum.GenSealedEnum
 import kotlinx.collections.immutable.toImmutableList
-import kotlinx.coroutines.launch
 
 interface CellShapeConfigUiInjectEntryPoint :
     ComposeLifePreferencesProvider
@@ -125,165 +102,6 @@ fun CellShapeConfigUi(
                     onSessionValueChange = currentShapeConfigUiState::onCornerFractionSessionValueChange,
                     valueRange = 0f..0.5f,
                     sliderBijection = Float.IdentitySliderBijection,
-                )
-            }
-        }
-    }
-}
-
-sealed interface ShapeDropdownOption : DropdownOption {
-    data object RoundRectangle : ShapeDropdownOption {
-        override val displayText: ParameterizedString = Strings.RoundRectangle
-    }
-
-    @GenSealedEnum
-    companion object
-}
-
-expect val ShapeDropdownOption.Companion._values: List<ShapeDropdownOption>
-
-interface CellShapeConfigUiState {
-    val currentShapeDropdownOption: ShapeDropdownOption
-
-    val currentShapeConfigUiState: CurrentShapeConfigUiState
-
-    fun setCurrentShapeType(option: ShapeDropdownOption)
-}
-
-sealed interface CurrentShapeConfigUiState {
-
-    interface RoundRectangleConfigUi : CurrentShapeConfigUiState {
-        val sizeFractionSessionValue: SessionValue<Float>
-
-        val cornerFractionSessionValue: SessionValue<Float>
-
-        fun onSizeFractionSessionValueChange(value: SessionValue<Float>)
-
-        fun onCornerFractionSessionValueChange(value: SessionValue<Float>)
-    }
-}
-
-context(ComposeLifePreferencesProvider, LoadedComposeLifePreferencesProvider)
-@Suppress("LongMethod")
-@Composable
-fun rememberCellShapeConfigUiState(): CellShapeConfigUiState {
-    val currentShapeType: CurrentShapeType = preferences.currentShapeType
-    val coroutineScope = rememberCoroutineScope()
-
-    val currentShapeConfigUiState = when (currentShapeType) {
-        is CurrentShapeType.RoundRectangle -> {
-            val roundRectangleSessionValueHolder = rememberSessionValueHolder<CurrentShape.RoundRectangle>(
-                upstreamSessionValue = preferences.roundRectangleSessionValue,
-                setUpstreamSessionValue = { expected, newValue ->
-                    coroutineScope.launch {
-                        composeLifePreferences.setRoundRectangleConfig(expected, newValue)
-                    }
-                },
-                valueSaver = listSaver(
-                    save = {
-                        listOf(it.sizeFraction, it.cornerFraction)
-                    },
-                    restore = {
-                        CurrentShape.RoundRectangle(
-                            it[0],
-                            it[1],
-                        )
-                    },
-                ),
-            )
-
-            val localSessionId = roundRectangleSessionValueHolder.info.localSessionId
-
-            val initialSizeFraction = remember(localSessionId) {
-                roundRectangleSessionValueHolder.sessionValue.value.sizeFraction
-            }
-            val initialCornerFraction = remember(localSessionId) {
-                roundRectangleSessionValueHolder.sessionValue.value.cornerFraction
-            }
-            var sizeFraction by remember(localSessionId) {
-                mutableFloatStateOf(initialSizeFraction)
-            }
-            var cornerFraction by remember(localSessionId) {
-                mutableFloatStateOf(initialCornerFraction)
-            }
-
-            var sizeFractionSessionId by key(localSessionId) {
-                rememberSaveable(stateSaver = uuidSaver) {
-                    mutableStateOf(uuid4())
-                }
-            }
-            var sizeFractionValueId by key(localSessionId) {
-                rememberSaveable(stateSaver = uuidSaver) {
-                    mutableStateOf(uuid4())
-                }
-            }
-            var cornerFractionSessionId by key(localSessionId) {
-                rememberSaveable(stateSaver = uuidSaver) {
-                    mutableStateOf(uuid4())
-                }
-            }
-            var cornerFractionValueId by key(localSessionId) {
-                rememberSaveable(stateSaver = uuidSaver) {
-                    mutableStateOf(uuid4())
-                }
-            }
-
-            object : CurrentShapeConfigUiState.RoundRectangleConfigUi {
-                override val sizeFractionSessionValue: SessionValue<Float> get() =
-                    SessionValue(
-                        sessionId = sizeFractionSessionId,
-                        valueId = sizeFractionValueId,
-                        value = sizeFraction,
-                    )
-
-                override val cornerFractionSessionValue: SessionValue<Float> get() =
-                    SessionValue(
-                        sessionId = cornerFractionSessionId,
-                        valueId = cornerFractionValueId,
-                        value = cornerFraction,
-                    )
-
-                override fun onSizeFractionSessionValueChange(value: SessionValue<Float>) {
-                    sizeFractionSessionId = value.sessionId
-                    sizeFractionValueId = value.valueId
-                    sizeFraction = value.value
-                    launchRoundRectangleConfigUpdate()
-                }
-
-                override fun onCornerFractionSessionValueChange(value: SessionValue<Float>) {
-                    cornerFractionSessionId = value.sessionId
-                    cornerFractionValueId = value.valueId
-                    cornerFraction = value.value
-                    launchRoundRectangleConfigUpdate()
-                }
-
-                private fun launchRoundRectangleConfigUpdate() {
-                    roundRectangleSessionValueHolder.setValue(
-                        CurrentShape.RoundRectangle(
-                            sizeFraction,
-                            cornerFraction,
-                        ),
-                    )
-                }
-            }
-        }
-    }
-
-    return object : CellShapeConfigUiState {
-        override val currentShapeDropdownOption: ShapeDropdownOption
-            get() = when (currentShapeType as CurrentShapeType) {
-                is CurrentShapeType.RoundRectangle -> ShapeDropdownOption.RoundRectangle
-            }
-
-        override val currentShapeConfigUiState: CurrentShapeConfigUiState
-            get() = currentShapeConfigUiState
-
-        override fun setCurrentShapeType(option: ShapeDropdownOption) {
-            coroutineScope.launch {
-                composeLifePreferences.setCurrentShapeType(
-                    when (option) {
-                        ShapeDropdownOption.RoundRectangle -> CurrentShapeType.RoundRectangle
-                    },
                 )
             }
         }
