+++ ui-common/src/jbTest/kotlin/com/alexvanyo/composelife/ui/util/SealedEnumSaverTests.kt
